dm: i2c: initial add i2c DM driver [1/1]
PD#162821

Problem:
Add i2c DM driver

Solution:
1.Refered to uboot upstream uboot driver, Porting
to meson SoCs.
2.i2c mw/md/dev/speed commands work.

Verify:
axg_s400_v3

+++ lib/fdtdec.c
@@ -110,7 +110,7 @@ fdt_addr_t fdtdec_get_addr_size(const void *blob, int node,
 					sizeof(fdt_addr_t));
 			*sizep = fdt_size_to_cpu(*size);
 			debug("addr=%08lx, size=%08x\n",
-			      (ulong)addr, *sizep);
+			      (ulong)addr, (unsigned int)*sizep);
 		} else {
 			debug("%08lx\n", (ulong)addr);
 		}
